feat: property select dimension
