+++ apps/cloud/src/app/@core/utils.ts
@@ -1,6 +1,7 @@
-import { Observable, firstValueFrom } from "rxjs"
-import { ToastrService } from "./services"
-import { getErrorMessage } from "./types"
+import { EntityType, getEntityDimensions, getEntityMeasures } from '@metad/ocap-core'
+import { firstValueFrom, Observable } from 'rxjs'
+import { ToastrService } from './services'
+import { getErrorMessage } from './types'
 
 export function requestFullscreen(document) {
   if (document.documentElement.requestFullscreen) {
@@ -38,7 +39,30 @@ export function exitFullscreen(document) {
 export async function tryHttp<T>(request: Observable<T>, toastrService?: ToastrService) {
   try {
     return await firstValueFrom(request)
-  } catch(err) {
+  } catch (err) {
     toastrService?.error(getErrorMessage(err))
   }
+}
+
+export function calcEntityTypePrompt(entityType: EntityType) {
+  return JSON.stringify({
+    name: entityType.name,
+    caption: entityType.caption,
+    dimensions: getEntityDimensions(entityType).map((dimension) => ({
+      name: dimension.name,
+      caption: dimension.caption,
+      hierarchies: dimension.hierarchies?.map((item) => ({
+        name: item.name,
+        caption: item.caption,
+        levels: item.levels?.map((item) => ({
+          name: item.name,
+          caption: item.caption
+        }))
+      }))
+    })),
+    measures: getEntityMeasures(entityType).map((item) => ({
+      name: item.name,
+      caption: item.caption
+    }))
+  })
 }
